+++ app/src/main/java/de/triology/cas/services/dogu/CesDoguServiceFactory.java
@@ -12,17 +12,19 @@ public class CesDoguServiceFactory implements ICesServiceFactory {
     public RegexRegisteredService createCASService(long id, String fqdn){
         RegexRegisteredService casService = new RegexRegisteredService();
         casService.setId(id);
-        casService.setServiceId("https://" + fqdn + "/cas/.*");
+        casService.setServiceId("https://" + generateServiceIdFqdnRegex(fqdn) + "/cas/.*");
         casService.setName(CesDoguServiceFactory.class.getSimpleName() + " " + SERVICE_CAS_IDENTIFIER);
         return casService;
     }
 
     @Override
     public RegexRegisteredService createNewService(long id, String fqdn, URI casLogoutURI, CesServiceData serviceData) throws CesServiceCreationException {
+        String fqdnRegex = generateServiceIdFqdnRegex(fqdn);
+
         RegexRegisteredService service = new RegexRegisteredService();
         service.setId(id);
 
-        String serviceId = "https://" + fqdn + "(:443)?/" + serviceData.getName() + "(/.*)?";
+        String serviceId = "https://" + fqdnRegex + "(:443)?/" + serviceData.getName() + "(/.*)?";
         service.setServiceId(serviceId);
         service.setName(serviceData.getIdentifier());
 
@@ -32,5 +34,13 @@ public class CesDoguServiceFactory implements ICesServiceFactory {
         }
 
         return service;
+    }
+
+    public static String generateServiceIdFqdnRegex(String fqdn) {
+        if (fqdn == null) {
+            return "";
+        }
+
+        return "((?i)" + fqdn.replace(".", "\\.") + ")";
     }
 }